fix: humble build fix

+++ oxidros/src/msg.rs
@@ -329,7 +329,7 @@ impl<const N: usize> RosString<N> {
         if self.0.data.is_null() {
             &[]
         } else {
-            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size) };
+            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size as usize) };
             s
         }
     }
@@ -338,7 +338,7 @@ impl<const N: usize> RosString<N> {
         if self.0.data.is_null() {
             &mut []
         } else {
-            let s = unsafe { std::slice::from_raw_parts_mut(self.0.data, self.0.size) };
+            let s = unsafe { std::slice::from_raw_parts_mut(self.0.data, self.0.size as usize) };
             s
         }
     }
@@ -411,7 +411,7 @@ impl<const STRLEN: usize, const SEQLEN: usize> RosStringSeq<STRLEN, SEQLEN> {
         if self.0.data.is_null() {
             &[]
         } else {
-            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size) };
+            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size as usize) };
             unsafe { transmute::<&[rosidl_runtime_c__String], &[RosString<STRLEN>]>(s) }
         }
     }
